refactor(upstream): extract pool/testsupport.rs shared test fixtures (lab-kvji.12.5 step 8)
Move shared #[cfg(test)] fixtures + mock servers (test_upstream_config, named_*,
test_tool*, StaticCatalogServer, SlowResponseServer, static_catalog_pool*,
slow_response_pool) into pool/testsupport.rs as pub(super). pool.rs test mod and
child test mods consume via use super[::super]::testsupport::*. Relocate the
deferred failed_in_process_entry_from_existing test into entries.rs (now that the
test_upstream_tools fixture is shared). pub(super) FLIP: with_request_timeout
(8th flip, not in plan §2.1 — testsupport calls it cross-module). Behavior
unchanged. Workspace just test: 1614 green.

Bead: lab-kvji.12.5
